// src/lib/types.ts
export type GoalType = 'college' | 'general' | 'neet' | 'jee' | 'school-boards' | 'ias' | 'school-prep' | 'olympiad' | 'govt-exams' | 'entrance-exam' | 'gate' | 'finance';

export interface UserGoal {
  type: GoalType;
  country: string; 
  university?: string;
  program?: string;
  branch?: string;
  semester?: string;
  board?: 'CBSE' | 'ICSE' | 'International';
  class?: string;
}